pyproject.toml
[tool.poetry]
name = "auth-telegram"
version = "0.1.0"
description = ""
authors = ["vlad-terehoff <tereh_off95@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
aiogram = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
